Get names of legislators that represent each attendee
The names of the legislators are acquired through the use of Google's
Civic Information API.

Because of the way that the API handles errors, the tutorial I am
following no longer works as intended in handling the records that are
missing data. Due to this, it was necessary to redirect the output
stream to a new StringIO object before the API calls are made, and then
substitute back the default output stream after the calls are made,
which makes it so that the errors are no longer output to the console.

This solution was provided by a user in The Odin Project's discord server:
https://discord.com/channels/505093832157691914/543074220691947531/1008539799549919314